Add vector data type
