+++ document-legacy/src/layers/mod.rs
@@ -25,6 +25,13 @@ pub mod layer_info;
 pub mod nodegraph_layer;
 /// Contains the [ShapeLayer](shape_layer::ShapeLayer) type, a generic SVG element defined using Bezier paths.
 pub mod shape_layer;
-pub mod style;
 /// Contains the [TextLayer](text_layer::TextLayer) type.
 pub mod text_layer;
+
+mod render_data;
+pub use render_data::RenderData;
+
+pub mod style {
+	pub use super::RenderData;
+	pub use graphene_core::vector::style::*;
+}
