+++ editor/Cargo.toml
@@ -23,8 +23,6 @@ serde_json = { version = "1.0" }
 graphite-proc-macros = { path = "../proc-macros" }
 bezier-rs = { path = "../libraries/bezier-rs" }
 glam = { version="0.22", features = ["serde"] }
-rand_chacha = "0.3.1"
-spin = "0.9.2"
 kurbo = { git = "https://github.com/linebender/kurbo.git", features = [
 	"serde",
 ] }
