+++ node-graph/gcore/Cargo.toml
@@ -33,6 +33,8 @@ kurbo = { git = "https://github.com/linebender/kurbo.git", features = [
 	"serde",
 ], optional = true }
 glam = { version = "^0.22", default-features = false, features = ["scalar-math", "libm"]}
+rand_chacha = "0.3.1"
+spin = "0.9.2"
 node-macro = {path = "../node-macro"}
 specta.workspace = true
 once_cell = { version = "1.17.0", default-features = false }
