+++ node-graph/gcore/src/consts.rs
@@ -1,0 +1,5 @@
+use crate::raster::Color;
+
+// RENDERING
+pub const LAYER_OUTLINE_STROKE_COLOR: Color = Color::BLACK;
+pub const LAYER_OUTLINE_STROKE_WEIGHT: f64 = 1.;
