+++ node-graph/gcore/src/raster/color.rs
@@ -28,7 +28,7 @@ pub struct Color {
 	alpha: f32,
 }
 
-#[allow(clippy::derive_hash_xor_eq)]
+#[allow(clippy::derived_hash_with_manual_eq)]
 impl Hash for Color {
 	fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
 		self.red.to_bits().hash(state);
