+++ node-graph/gcore/src/vector/mod.rs
@@ -1,6 +1,16 @@
 pub mod consts;
 pub mod generator_nodes;
-pub mod id_vec;
 pub mod manipulator_group;
 pub mod manipulator_point;
+
+pub mod style;
+pub use style::PathStyle;
+
 pub mod subpath;
+pub use subpath::Subpath;
+
+mod vector_data;
+pub use vector_data::VectorData;
+
+mod id_vec;
+pub use id_vec::IdBackedVec;
